auth_service/requirements.txt
# Core Django framework
django==4.2.16

# Django REST Framework for building REST APIs
djangorestframework==3.15.2

# JWT authentication for DRF
djangorestframework-simplejwt==5.3.1

# PostgreSQL adapter for Django
psycopg2-binary==2.9.9

# Redis integration for caching and Celery
django-redis==5.4.0
redis==5.0.8


# RabbitMQ client for event-driven messaging
pika==1.3.2

# For environment variable management
python-dotenv==1.0.1

celery[redis,email]==5.4.0
